+++ beets/ui/commands/import_/session.py
@@ -97,7 +97,12 @@ class TerminalImportSession(importer.ImportSession):
             # or, basic choices that require no more action here.
             if isinstance(choice, AlbumMatch) or (
                 isinstance(choice, importer.Action)
-                and choice in (importer.Action.SKIP, importer.Action.ASIS)
+                and choice
+                in (
+                    importer.Action.SKIP,
+                    importer.Action.ASIS,
+                    importer.Action.RESCAN,
+                )
             ):
                 # Pass selection to main control flow.
                 return choice
@@ -249,6 +254,14 @@ class TerminalImportSession(importer.ImportSession):
                     "g", "Group albums", lambda s, t: importer.Action.ALBUMS
                 ),
             ]
+            if task.toppath:
+                choices.append(
+                    PromptChoice(
+                        "r",
+                        "Rescan directory",
+                        lambda s, t: importer.Action.RESCAN,
+                    )
+                )
         choices += [
             # TODO: introduce beets.autotag.Candidates to remove these ignores
             #  context: Candidates is a Sequence which will be updated in place